chore: Refactor Field Data.
