+++ src/ansys/fluent/core/session_solver.py
@@ -35,7 +35,7 @@ from ansys.fluent.core.exceptions import BetaFeaturesNotEnabled
 from ansys.fluent.core.module_config import config
 from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
 from ansys.fluent.core.services import MonitorsServiceV0, service_creator
-from ansys.fluent.core.services.field_data import ZoneInfo, ZoneType
+from ansys.fluent.core.services._field_data import ZoneInfo, ZoneType
 from ansys.fluent.core.services.monitor_v1 import MonitorsService
 from ansys.fluent.core.services.scheme_interpreter import SchemeInterpreter
 from ansys.fluent.core.services.solution_variables import (
