+++ tests/test_field_data.py
@@ -41,7 +41,7 @@ from ansys.fluent.core.field_data_interfaces import (
     FieldUnavailable,
     _Fields,
 )
-from ansys.fluent.core.services.field_data import (
+from ansys.fluent.core.services._field_data import (
     CellElementType,
     ZoneType,
 )
